requirements.txt
pyTelegramBotAPI
requests
APScheduler
python-dotenv
